Finished Authentication

+++ frontend/src/components/RecipeDetails.js
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 // NOTE: RECIPE DETAILS ---------------------------------------------------------------------------
 const RecipeDetails = ({ recipe }) => {
@@ -6,14 +7,22 @@ const RecipeDetails = ({ recipe }) => {
     // NOTE: URL 
     const urlName = recipe._id
 
+    // VARIABLE TO STORE LOGGED IN USER
+    const { user } = useAuthContext()
+
     // NOTE: HANDLING DELETE ----------------------------------------------------------------------
     const handleDelete = async () => {
-        const response = await fetch(`http://localhost:4000/recipes/` + recipe._id, {
-            method: 'DELETE',
-        });
+        if (user) {
+            const response = await fetch(`http://localhost:4000/recipes/` + recipe._id, {
+                headers: {
+                    'Authorization': `Bearer ${user.token}`
+                },
+                method: 'DELETE',
+            });
 
-        if (response.ok) {
-            console.log("Deleted: " + recipe._id)
+            if (response.ok) {
+                console.log("Deleted: " + recipe._id)
+            }
         }
     }
